(fix): fix insecure login and insecure second layer file naming, made functions for plugins
